last kernel 2026 final changes
git-svn-id: https://svn.code.sf.net/p/freedos/svn/kernel/trunk@349 6ac86273-5f31-0410-b378-82cca8765d1b

+++ sys/fdkrncfg.c
@@ -24,8 +24,8 @@ char KERNEL[] = "KERNEL.SYS";
 #include <fcntl.h>
 
 #include "portab.h"
-extern WORD CDECL printf(CONST BYTE * fmt, ...);
-extern WORD CDECL sprintf(BYTE * buff, CONST BYTE * fmt, ...);
+extern int CDECL printf(CONST BYTE * fmt, ...);
+extern int CDECL sprintf(BYTE * buff, CONST BYTE * fmt, ...);
 
 #ifdef __WATCOMC__
 #define close _dos_close
